+++ src/main/java/org/folio/invoices/utils/UploadHelper.java
@@ -2,12 +2,13 @@ package org.folio.invoices.utils;
 
 import java.util.concurrent.CompletableFuture;
 
+import org.folio.rest.acq.model.BatchVoucher;
+
 public interface UploadHelper {
 
   public CompletableFuture<String> login(String username, String password);
 
   public CompletableFuture<String> logout();
 
-  // upload method to be added later (MODINVOICE-133)
-
+  public CompletableFuture<String> upload(String filename, BatchVoucher batchVoucher);
 }
